Image uploader refactor

+++ src/Controller/DashboardController.php
@@ -7,9 +7,11 @@ use App\Form\ChangePasswordFormType;
 use App\Form\DeleteAccountFormType;
 use App\Form\ImageFormType;
 use App\Form\UserFormType;
+use App\Services\ImageUploader;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Bundle\SecurityBundle\Security;
+use Symfony\Component\HttpFoundation\File\Exception\FileException;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -25,39 +27,48 @@ class DashboardController extends AbstractController
     }
 
     #[Route('/dashboard/profile', name: 'app_profile')]
-    public function profile(Request $request, EntityManagerInterface $entityManager,
-    Security $security): Response
+    public function profile(Request  $request, EntityManagerInterface $entityManager,
+                            Security $security, ImageUploader $imageUploader): Response
     {
         // change image
         $image = new Image();
-        $imageForm = $this -> createForm(ImageFormType::class, $image);
+        $imageForm = $this->createForm(ImageFormType::class, $image);
         $imageForm->handleRequest($request);
         $user = $this->getUser();
-        if($imageForm->isSubmitted() && $imageForm->isValid()){
-//            $image = $imageForm -> getData();
-            $image->setPath($imageForm->get('imageFile')->getData()
-                ->getClientOriginalName());
-            if ($user->getImage()){
-                $oldImage = $entityManager->getRepository(Image::class)->find($user->getImages()->getId());
-                $entityManager -> remove($oldImage);
+        if ($imageForm->isSubmitted() && $imageForm->isValid()) {
+            $imageFile = $imageForm->get('imageFile')->getData();
+            if ($imageFile) {
+                if ($user->getImage()?->getPath()) {
+                    unlink($this->getParameter('images_directory') . '/' . $user->getImage()
+                            ->getPath());
+                }
+                $newFilename = $imageUploader->upload($imageFile);
+                $image->setPath($newFilename);
+                if ($user->getImage()) {
+                    $oldImage = $entityManager->getRepository(Image::class)->find($user->getImage()->getId());
+                    $entityManager->remove($oldImage);
+
+                }
+
+                $user->setImage($image);
+                $entityManager->persist($image);
+                $entityManager->persist($user);
+                $entityManager->flush();
+                $this->addFlash(
+                    'status-image',
+                    'image-updated'
+                );
+
             }
-            $user->setImage($image);
-            $entityManager->persist($image);
-            $entityManager->persist($user);
-            $entityManager->flush();
-            $this->addFlash(
-                'status-image',
-                'image-updated'
-            );
             return $this->redirectToRoute('app_profile');
         }
 
         // change user email, name
-        $user= $this->getUser();
+        $user = $this->getUser();
         $userForm = $this->createForm(UserFormType::class, $user);
         $userForm->handleRequest($request);
 
-        if ($userForm->isSubmitted() && $userForm->isValid()){
+        if ($userForm->isSubmitted() && $userForm->isValid()) {
 //            $user = $userForm->getData();
             $entityManager->persist($user);
             $entityManager->flush();
@@ -71,7 +82,7 @@ class DashboardController extends AbstractController
         // change password
         $passwordForm = $this->createForm(ChangePasswordFormType::class, $user);
         $passwordForm->handleRequest($request);
-        if ($passwordForm->isSubmitted() && $passwordForm->isValid()){
+        if ($passwordForm->isSubmitted() && $passwordForm->isValid()) {
 //            $user = $passwordForm->getData();
             $entityManager->persist($user);
             $entityManager->flush();
@@ -84,10 +95,10 @@ class DashboardController extends AbstractController
 
         // delete account
         $deleteAccountForm = $this->createForm(DeleteAccountFormType::class, $user);
-        $deleteAccountForm -> handleRequest($request);
-        if ($deleteAccountForm->isSubmitted() && $deleteAccountForm->isValid()){
+        $deleteAccountForm->handleRequest($request);
+        if ($deleteAccountForm->isSubmitted() && $deleteAccountForm->isValid()) {
 //            $user = $deleteAccountForm->getData();
-            $security -> logout(false);
+            $security->logout(false);
             $entityManager->remove($user);
             $entityManager->flush();
             $request->getSession()->invalidate();
